pyproject.toml
[project]
name = "kmd-nexus-client"
version = "0.1.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.4.0",
    "httpx>=0.28.1",
    "pytest>=8.3.4",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "ruff>=0.12.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = [
    "/kmd_nexus_client",
    "/README.md",
    "/tests",
]
